Commenting of variables;  Added option to not delete m4v if itunes does not organize library folders

+++ sickbeard_to_itunes.rb
@@ -1,11 +1,28 @@
+# Path to your SickBeard library
 sb_lib = "/Volumes/500GB/Sick\\ Beard"
+
+# Path to your iTunes library
 it_lib = "/Volumes/2TB/iTunes"
+
+# iTunes copies imported files
+organized_itunes = true
+
+# Where you have HandBrakeCLI installed
 pathToHB = "/Applications/"
+
+# Temporary path your m4v files should end up before being imported to iTunes
 outPath = "/Users/zach/Movies/"
+
+# Target quality for HandBrake
 preset = "AppleTV 2"
+
+# Other options for HandBrake
 options = ""
+
+# Unix priority 19 will give up priority so your computer isn't _as_ slow
 nice = "/usr/bin/nice -n 19"
 
+# The code, no need to edit any of this unless you know what you're doing!
 def get_filename(path)
     path.split("/").last.split(".").first.chomp
 end
@@ -36,7 +53,6 @@ while 1 == 1
 
   `find #{sb_lib} -iname *\.mkv -o -iname *\.avi -o -iname *\.wmv`.each do |f|
     filename = get_filename(f)
-    # First we check for old style file names
     info = create_show_details(filename)
     #  This checks for new style names and old style names
     i = `find #{it_lib} -iname "*#{filename.gsub(/\W/,"*")}*"`.chomp.size + `find #{it_lib} -iname "*#{get_title(info)}*"`.chomp.size
@@ -53,13 +69,26 @@ while 1 == 1
       outPathWName.gsub!(/\'/,"\\'")
       track_id = `/usr/bin/osascript -e 'tell application \"iTunes\" to add POSIX file \"#{outPathWName}\"'`.chomp.split(" ")[3]
       info = create_show_details(filename)
+      
+      # Set the episode's title
       system "/usr/bin/osascript -e 'tell application \"iTunes\" to set name of track id #{track_id} to \"#{get_title(info)}\"'"
+      
+      # Change the Media Kind to TV Show
       system "/usr/bin/osascript -e 'tell application \"iTunes\" to set video kind of track id #{track_id} to tv show'"
+      
+      # Set the season number
       system "/usr/bin/osascript -e 'tell application \"iTunes\" to set season number of track id #{track_id} to #{get_season_number(info)}'"
+      
+      # Set the episode number
       system "/usr/bin/osascript -e 'tell application \"iTunes\" to set episode number of track id #{track_id} to #{get_episode_number(info)}'"
+      
+      # Set the name of the show
       system "/usr/bin/osascript -e 'tell application \"iTunes\" to set show of track id #{track_id} to \"#{get_show_name(info)}\"'"
+      
       sleep 10
-      if track_id.to_i > 0
+      
+      # If the initial file addition was successful, itunes copied everything and we can delete our temporary file
+      if organized_itunes and track_id.to_i > 0
         File.delete(outPathWName)
       end
   end
